Start adding contributed logic for compute user and group creation
NickyRuggs contributed code to dumps the contents of the local user and
groups created by these scripts into /etc/passwd and /etc/group output
so that service accounts on compute will match those on the Isilon.
While not technically necessary it will be a best practice for these to
match going forward for sanity.